refactor(domain): move MustParsePath to testing.go
Move test-only function MustParsePath from path.go to dedicated
testing.go file. This clearly marks it as test-only code and prevents
accidental use in production.

Function still available in domain package for test usage.

Part of Phase 24.8 Technical Debt Cleanup.
