+++ internal/domain/path.go
@@ -132,13 +132,3 @@ func clean(path string) string {
 	}
 	return cleaned
 }
-
-// MustParsePath creates a FilePath from a string, panicking on error.
-// This function is intended for use in tests and should not be used in production code.
-func MustParsePath(s string) FilePath {
-	result := NewFilePath(s)
-	if result.IsErr() {
-		panic(result.UnwrapErr())
-	}
-	return result.Unwrap()
-}
